Dockerizing...

+++ prepare.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 echo "creating js folders..."
 mkdir -p vendor/js && mkdir -p vendor/css && mkdir -p vendor/fonts 
 echo "copying required stuff..."
@@ -7,6 +8,12 @@ cp -a vendor/components/jquery/jquery.min.js vendor/js/jquery.min.js &&
 cp -a vendor/twbs/bootstrap/dist/css/. vendor/css/ && 
 cp -a vendor/twbs/bootstrap/dist/js/. vendor/js/ && 
 cp -a vendor/twbs/bootstrap/dist/fonts/. vendor/fonts/
-echo "removing not required stuff..."
-rm -Rf vendor/components && rm -Rf vendor/twbs && rm -Rf .git/ && rm -Rf .win && rm nginx.conf
-echo "ready to distribute"
+if [ -z "${DEVELOPMENT}" ]
+then
+   echo "removing not required stuff..."
+   rm -Rf vendor/components && rm -Rf vendor/twbs  && rm -Rf .git/  && rm -Rf .win && rm nginx.conf
+   echo "ready to distribute"
+else
+    echo "Running in DEV mode!"
+    php-fpm
+fi
